Fix clippy warnings for Rust 1.95.

+++ lgbm/tests/booster.rs
@@ -232,7 +232,7 @@ fn update_one_iter_custom() -> Result<()> {
         let p = b.get_predict(0)?;
         for i in 0..train_label.len() {
             let p = (p[i] as f32).clamp(0.0001, 0.9999);
-            let g = p as f32 - train_label[i];
+            let g = p - train_label[i];
             grad.push(g);
             hess.push(p * (1.0 - p));
         }
